[JENKINS-18110] Look for a project by it's identifier and then by name
Redmine project should be mainly looked by it's identifier, because it
doesn't contain spaces and its case insensitive.

With previous behaviour, when project identifier was provided, metrics
aggregation worked only for project with the same name as identifier.
Search by name is left for backward compatibility, because when someone
wanted metrics agregation, then configuration still contains project
name (which might contain spaces), so we cant't break this for anyone.

The class fields were not refactored to "identifier" to simplify the fix
- the explanation has been added as a Javadoc comment and in jelly views